Cleaning up some stk documentation
I've tried to unify some of the doxygen output to reduce clutter and
missing content.  I added some content for stk_rebalance and have
some more to do...

Build/Test Cases Summary
Enabled Packages: STK
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=8,notpassed=0 (6.73 min)
1) SERIAL_RELEASE => passed: passed=6,notpassed=0 (3.10 min)

+++ packages/stk/stk_rebalance/stk_rebalance/ZoltanPartition.cpp
@@ -846,12 +846,12 @@ int  Zoltan::evaluate( int    print_stats,
   ZOLTAN_GRAPH_EVAL   graph = {{0}};
   if (Zoltan_LB_Eval_Balance( m_zoltan_id_, print_stats, &eval)) ierr = 1;
   if (Zoltan_LB_Eval_Graph( m_zoltan_id_, print_stats, &graph) ) ierr = 1;
-  *nobj         = eval.nobj[0];
-  *obj_wgt      = eval.obj_wgt[0];
-  *ncuts        = graph.cuts[0];
-  *cut_wgt      = graph.cut_wgt[0];
-  *nboundary    = graph.num_boundary[0];
-  *nadj         = graph.nnborparts[0];
+  *nobj         = (int)eval.nobj[0];
+  *obj_wgt      =      eval.obj_wgt[0];
+  *ncuts        = (int)graph.cuts[0];
+  *cut_wgt      =      graph.cut_wgt[0];
+  *nboundary    = (int)graph.num_boundary[0];
+  *nadj         = (int)graph.nnborparts[0];
 
   return ierr;
